third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/dom/subtree-modified-attributes-expected.txt

CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
DOMSubtreeModified should fire when attributes are added or removed, but not modified (see bug 81141)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expected is true
PASS expected is true
PASS expected is true
PASS expected is true
PASS successfullyParsed is true

TEST COMPLETE
